// src/app/primary/utils/enums.ts
export enum Category {
    OperatingSystem = 'OperatingSystem',
    ComputerNetworking = 'ComputerNetworking',
    Cryptography = 'Cryptography',
    InQuizitive = 'InQuizitive',
    FinView = 'FinView',
    OmniSphere = 'OmniSphere',
  }

  export enum Subcategory {
    Medium = 'Medium',
    Project = 'Project',
  }